+++ ktor/server/src/jvmMain/kotlin/dev/inmo/micro_utils/ktor/server/FlowsWebsocket.kt
@@ -4,23 +4,24 @@ import dev.inmo.micro_utils.coroutines.safely
 import dev.inmo.micro_utils.ktor.common.*
 import io.ktor.application.featureOrNull
 import io.ktor.application.install
+import io.ktor.http.URLProtocol
 import io.ktor.http.cio.websocket.*
 import io.ktor.routing.Route
 import io.ktor.routing.application
-import io.ktor.websocket.WebSocketServerSession
-import io.ktor.websocket.webSocket
+import io.ktor.websocket.*
 import kotlinx.coroutines.flow.Flow
 import kotlinx.serialization.SerializationStrategy
 
 fun <T> Route.includeWebsocketHandling(
     suburl: String,
     flow: Flow<T>,
+    protocol: URLProtocol = URLProtocol.WS,
     converter: suspend WebSocketServerSession.(T) -> StandardKtorSerialInputData?
 ) {
     application.apply {
         featureOrNull(io.ktor.websocket.WebSockets) ?: install(io.ktor.websocket.WebSockets)
     }
-    webSocket(suburl) {
+    webSocket(suburl, protocol.name) {
         safely {
             flow.collect {
                 converter(it) ?.let { data ->
@@ -36,10 +37,12 @@ fun <T> Route.includeWebsocketHandling(
     flow: Flow<T>,
     serializer: SerializationStrategy<T>,
     serialFormat: StandardKtorSerialFormat = standardKtorSerialFormat,
+    protocol: URLProtocol = URLProtocol.WS,
     filter: (suspend WebSocketServerSession.(T) -> Boolean)? = null
 ) = includeWebsocketHandling(
     suburl,
     flow,
+    protocol,
     converter = if (filter == null) {
         {
             serialFormat.encodeDefault(serializer, it)